OM-37954 RB-26088 Refactor CostDTO for cloud integration in XL

+++ com.vmturbo.analysis/platform.analysis/src/main/java/com/vmturbo/platform/analysis/economy/BalanceAccount.java
@@ -9,9 +9,9 @@ public class BalanceAccount {
 
     private double spent_;
     private double budget_;
-    private int id_;
+    private long id_;
 
-    public BalanceAccount(double spent, double budget, int id) {
+    public BalanceAccount(double spent, double budget, long id) {
         spent_ = spent;
         budget_ = budget;
         id_ = id;
@@ -33,7 +33,7 @@ public class BalanceAccount {
         return budget_;
     }
 
-    public int getId() {
+    public long getId() {
         return id_;
     }
 }
